Remove associatedtype from ReusableViewProtocol
This isn't going to work, because then the protocol can only be used as a generic constraint, which we don't want.
Such a constraint would require that all cells be the same type, and we want to be able to have different cell types.

Plus, maybe cells shouldn't know their container anyway.