Fix bug in data source implementation
This was tricky. The "shuffle" and "reload" functions in the example app were not working. Cells were getting all mixed up and not updating correctly, resulting in many duplicates and other odd behavior.

The issue was that `_DiffableDataSource` was capturing `CollectionViewModel` (which is a struct!) in its cell provider and supplementary view provider.
Thus, the data source was never actually getting fully configured by each snapshot. The initial view model (captured in the closures) was always being used to configure cells.

The fix is to pass the closures into `_DiffableDataSource` from `CollectionViewDriver`, which is a reference type. The closures can safely capture `self` now (that is, the driver), which will always be the same reference.
The driver can then refer to the latest view model property it has.

Other changes:

This also changes the data source `applySnapshot` to accept the previous and current snapshots. This will be used in future work to determine item reloads/reconfigures.