Introduce `CellEventCoordinator`, remove direct view controller access
This replaces passing in the owning `UIViewController` for cell selection events with a protocol `CellEventCoordinator`.

This is more flexible for clients. You can now pass in any object you want. Perhaps that's some kind of custom `Navigator`, etc. that you use for routing in your app.

The goal here is to abstract away usage of `UIViewController` for clients who wish to do that. Otherwise, folks can have their `UIViewController` conform to `CellEventCoordinator` and proceed like before this change.

Another option is to implement cell selection in your view model directly and handle events _without_ using the coordinator.