Only reconfigure visible cells when applying snapshot
optimize the cells that we reconfigure.

previously, we were (almost) reconfiguring everything.

first, this only reloads what is visible. very important.

second, we also short-circuit our loop when either the `destinationCell` or the `sourceCell` do not exist. previously this would have resulted in unnecessary reconfigures because we might have been comparing to `!= nil`.